requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
groq==0.4.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
aiohttp==3.9.1
websockets==12.0
pydantic==2.5.3
python-multipart==0.0.6
requests==2.31.0
pillow==10.2.0
apscheduler==3.10.4
jinja2==3.1.3
pyjwt==2.8.0
passlib[bcrypt]==1.7.4

# Integraciones de Pago
mercadopago==2.2.1
paypalrestsdk==1.13.1
stripe==7.8.0

# Procesamiento Multimedia
openai==1.6.1
gtts==2.5.0
pytesseract==0.3.10
SpeechRecognition==3.10.1

# Email
python-email-validator==2.1.0
